_stream.py: Added new _reset() function.
Instead of calling self._scheduler.clear_queues(), add a focus point
for resetting state related to a scheduler run. I suspect that the
session element counts have been flawed in cases where we run the
scheduler multiple times but fail to clear that part of the state.

This should all around be a better approach for multiple scheduler runs.